refactor(artifact-provenance): extract producer capture
