+++ scripts/ci/module-impact.json
@@ -148,6 +148,7 @@
         "src/main/artifacts/provenance-canonical.ts",
         "src/main/artifacts/provenance-execution-evidence.ts",
         "src/main/artifacts/provenance-message-snapshot.ts",
+        "src/main/artifacts/provenance-producer-capture.ts",
         "src/main/artifacts/provenance-repository.ts",
         "src/main/artifacts/provenance-version-writer.ts"
       ],
